+++ packages/orq-rc/src/models/operations/generateconversationname.ts
@@ -251,7 +251,7 @@ export const GenerateConversationNameResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kfaqp7w0cjweb73nf6amphe5"),
+  _id: z.string().default("conv_01kfb0s05c58dmc8j7bpq83y4j"),
   kind: GenerateConversationNameKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),
@@ -283,7 +283,7 @@ export const GenerateConversationNameResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   GenerateConversationNameResponseBody
 > = z.object({
-  id: z.string().default("conv_01kfaqp7w0cjweb73nf6amphe5"),
+  id: z.string().default("conv_01kfb0s05c58dmc8j7bpq83y4j"),
   kind: GenerateConversationNameKind$outboundSchema,
   displayName: z.string(),
   createdAt: z.number(),