+++ packages/orq-rc/src/models/operations/retrievedatasource.ts
@@ -145,7 +145,7 @@ export const RetrieveDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KFAQP8AVFENG5ZBXXK4PFBRG"),
+  _id: z.string().default("01KFB0S0Z9C54XF6X3QVEZK5EC"),
   display_name: z.string(),
   description: z.string().optional(),
   status: RetrieveDatasourceStatus$inboundSchema,
@@ -188,7 +188,7 @@ export const RetrieveDatasourceResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   RetrieveDatasourceResponseBody
 > = z.object({
-  id: z.string().default("01KFAQP8AVFENG5ZBXXK4PFBRG"),
+  id: z.string().default("01KFB0S0Z9C54XF6X3QVEZK5EC"),
   displayName: z.string(),
   description: z.string().optional(),
   status: RetrieveDatasourceStatus$outboundSchema,