+++ packages/orq-rc/src/models/operations/updatedatasource.ts
@@ -202,7 +202,7 @@ export const UpdateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KFAQP8AZNEW8AHPX5FGTJB6W"),
+  _id: z.string().default("01KFB0S0ZDPA99XJ1RSMP4CC9T"),
   display_name: z.string(),
   description: z.string().optional(),
   status: UpdateDatasourceStatus$inboundSchema,
@@ -245,7 +245,7 @@ export const UpdateDatasourceResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   UpdateDatasourceResponseBody
 > = z.object({
-  id: z.string().default("01KFAQP8AZNEW8AHPX5FGTJB6W"),
+  id: z.string().default("01KFB0S0ZDPA99XJ1RSMP4CC9T"),
   displayName: z.string(),
   description: z.string().optional(),
   status: UpdateDatasourceStatus$outboundSchema,